% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FREDCPI.R
\docType{data}
\name{FREDCPI}
\alias{FREDCPI}
\title{Calculated US CPI Data Set Through Time}
\format{A data frame with 842 rows and 2 variables:
\describe{
  \item{DATE}{Date of the calculated CPI, from Jan. 1, 1947 until Feburary 1, 2017}
  \item{CPIAUCSL}{Estimated CPI with base year being the average of the price index for 1982-1984 seasonally adjusted}
  ...
}}
\source{
\url{https://fred.stlouisfed.org/series/CPIAUCSL}
}
\usage{
FREDCPI
}
\description{
This contains the Federal Reserve Bank of Saint Louis's estimated CPI for All Urban Consumers from Janurary 1947 until Feburary 2017
}
\keyword{datasets}
